implement some more tests fix bugs
